+++ package.json
@@ -5,7 +5,7 @@
   "main": "web.js",
   "_comment_1": "we should probably change the main script, it shouldn't start a server, says http://package.json.nodejitsu.com/ (hover over the main part on there)",
   "scripts": {
-    "start": "node web.js",
+    "start": "node lib/index.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
