+++ restart_node
@@ -1,2 +1,2 @@
-sudo kill $(pgrep -f "sudo PORT=80 node web.js");
-nohup sudo PORT=80 node web.js &
+sudo kill $(pgrep -f "sudo PORT=80 node lib/index.js");
+nohup sudo PORT=80 node lib/index.js &
