--- web.js
@@ -1,850 +1,0 @@
-/*
-	LiveChan is a live imageboard web application.
-    Copyright (C) 2014 LiveChan Team
-
-    This program is free software: you can redistribute it and/or modify
-    it under the terms of the GNU Affero General Public License as
-    published by the Free Software Foundation, either version 3 of the
-    License, or (at your option) any later version.
-
-    This program is distributed in the hope that it will be useful,
-    but WITHOUT ANY WARRANTY; without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-    GNU Affero General Public License for more details.
-
-    You should have received a copy of the GNU Affero General Public License
-    along with this program.  If not, see <http://www.gnu.org/licenses/>. 
-*/
-
-/* required modules */
-var express = require("express");
-var logfmt = require("logfmt");
-var http = require('http');
-var https = require('https');
-var crypto = require('crypto');
-var ipfilter = require('ipfilter');
-var captcha = require('captcha');
-var tripcode = require('tripcode');
-var fs = require('fs');
-var format = require('util').format;
-var mongoose = require('mongoose');
-var exec = require('child_process').exec;
-
-var useImageMagick = true;
-
-/* salt */
-var securetrip_salt = "AVEPwfpR4K8PXQaKa4PjXYMGktC2XY4Qt59ZnERsEt5PzAxhyL";
-
-fs.readFile('salt.txt', 'utf8', function (err, data) {
-    "use strict";
-    if (!err) {
-        securetrip_salt = data;
-    }
-});
-
-/* initialize app */
-var app = express();
-var port = process.env.PORT || 5000;
-var secure_port = 443;
-if (port != 80) secure_port = 5443;
-
-/* listen now */
-var server = http.createServer(app).listen(port, function () {
-    "use strict";
-    console.log('Express server listening on port %d in %s mode', server.address().port, app.settings.env);
-});
-
-var options = {
-  ca:   fs.readFileSync('sub.class1.server.ca.pem'),
-  key:  fs.readFileSync('ssl.key'),
-  cert: fs.readFileSync('ssl.crt')
-};
-
-var secure_server;
-try {
-	secure_server = https.createServer(options, app).listen(secure_port, function () {
-	    "use strict";
-	    console.log('Express secure_server listening on port %d in %s mode', secure_server.address().port, app.settings.env);
-	});
-} catch (e) {
-	console.log(e);
-	console.log("Running with insecure server");
-	secure_server = server;
-}
-
-var io = require('socket.io').listen(secure_server);
-
-/* set up db */
-mongoose.connect('mongodb://localhost/livechan_db');
-var Schema = mongoose.Schema;
-
-
-/* 5mb limit */
-app.use(express.limit('5mb'));
-
-/* for saving images */
-app.use(express.bodyParser({
-    uploadDir: 'public/tmp/uploads',
-    keepExtensions: true
-}));
-
-/* blocked nodes. this file can include more */
-fs.readFile('tor_list.txt', 'utf8', function (err, data) {
-    "use strict";
-    if (err) {
-        return;
-    }
-    var tor_list = data.split("\n"); // IP per line
-    app.use(ipfilter(tor_list));
-});
-
-/* logging only during development */
-if (port!=80)
-	app.use(logfmt.requestLogger());
-
-/* serve public data (/public/*) and get cookies */
-app.use(express.static(__dirname + '/public'));
-app.use(express.cookieParser());
-app.use(express.cookieSession({
-    secret: 'keyboard-cat'
-}));
-
-/* captcha security */
-app.use(captcha({
-    url: '/captcha.jpg',
-    color: '#465e67',
-    background: '#800000'
-}));
-
-/* stored data in memory */
-var count;
-var hash_list = [];
-var session_list = [];
-var ips = {};
-var boards = ['all', 'a', 'b', 'c', 'd', 'e', 'f', 'film', 'g', 'gif', 'h', 'hr', 'k', 'm', 'o', 'p', 'r', 's', 't', 'u', 'v', 'vg', 'vr', 'w', 'wg', 'i', 'ic', 'r9k', 's4s', 'cm', 'hm', 'lgbt', 'y', '3', 'adv', 'an', 'asp', 'cgl', 'ck', 'co', 'diy', 'fa', 'fit', 'gd', 'hc', 'int', 'jp', 'lit', 'mlp', 'mu', 'n', 'out', 'po', 'pol', 'sci', 'soc', 'sp', 'tg', 'toy', 'trv', 'tv', 'vp', 'waifu', 'wsg', 'x', 'dev', 'tech', 'prog','dogecoin','fedoracoin', 'coin', 'q'];
-
-/* database fields to transmit */
-var all_fields = 'chat name body convo convo_id count date trip';
-var board_fields = 'chat name body convo convo_id count date image image_filename image_filesize image_width image_height thumb trip';
-
-/* db schema */
-var chat_schema = new Schema({
-    convo: String,
-    convo_id: Number,
-    is_convo_op: Boolean,
-    body: String,
-    name: String,
-    count: {
-        type: Number,
-        unique: true
-    },
-    date: {
-        type: Date,
-        default: Date.now
-    },
-    ip: String,
-    chat: String,
-    image: String,
-    image_filename: String,
-    image_filesize: Number,
-    image_width: Number,
-    image_height: Number,
-    thumb: String,
-    trip: String
-});
-
-var user_schema = new Schema({
-    date: { /* user session expires every 24 hours */
-        type: Date,
-        default: Date.now,
-        expires: '24h'
-    },
-    ip: String,
-    session_key: String,
-    captcha: String,
-    last_post: {
-        type: Date,
-        default: Date.now
-    },
-    banned_rooms: Array /* list of rooms the user is banned from */
-});
-
-var chat_db = mongoose.model('chat_db', chat_schema);
-var user_db = mongoose.model('user_db', user_schema);
-
-/* set counter to newest post */
-
-chat_db.findOne().sort({
-    count: -1
-}).exec(function (e, d) {
-    "use strict";
-    if (d) {
-        count = d.count;
-    }
-});
-
-/* helper functions */
-
-function get_extension(filename) {
-    "use strict";
-    var i = filename.lastIndexOf('.');
-    return (i < 0) ? '' : filename.substr(i + 1).toLowerCase();
-}
-
-function invalid_extension(filename) {
-    "use strict";
-    var test = ['jpg', 'jpeg', 'png', 'gif', 'ogv', 'webm'].indexOf(get_extension(filename));
-    if (test > -1) {
-        return false;
-    }
-    return true;
-}
-
-function is_video(filename) {
-    "use strict";
-    return (['ogv', 'webm'].indexOf(get_extension(filename)) > -1);
-}
-
-function delete_posts(e, ds) {
-    "use strict";
-    if (e) {
-        console.log("chat find error", e);
-        return;
-    }
-    ds.forEach(function(d) {
-        if (d.image) fs.unlink(d.image);
-        if (d.thumb) fs.unlink(d.thumb);
-        d.remove(function(e2) {console.log("chat removal error",e2);});
-    });
-}
-
-function send_data(room, event, data, fields) {
-    "use strict";
-    var fields_array = fields.split(" ");
-    var data2 = {};
-    for (var key in data) {
-        if (fields_array.indexOf(key) > -1) {
-            data2[key] = data[key];
-        }
-    }
-    io.sockets.in(room).emit(event, data2);
-}
-
-function add_to_chat(data) {
-    "use strict";
-    send_data(data.chat, 'chat', data, board_fields);
-    send_data('all', 'chat', data, all_fields);
-
-    /* store in the db */
-    chat_db.update({count: data.count}, data, {upsert: true}, function(err) {
-        if (err) {
-            console.log("chat save error", err);
-            return;
-        }
-        chat_db.find({chat: data.chat, is_convo_op: false})
-            .sort({count: -1})
-            .skip(100)
-            .exec(delete_posts);
-        chat_db.find({chat: data.chat, is_convo_op: true})
-            .sort({count: -1})
-            .skip(100)
-            .exec(delete_posts);
-    });
-}
-
-function session_exists(session) {
-    "use strict";
-    if (user_db.count({
-        session_key: session
-    })) {
-        return true;
-    }
-    return false;
-}
-
-/* check_ip_validity:
-	- checks if ip is banned
-	- checks session validity
-	- checks cool down 
-	calls format_chat(req, res, next, add_to_chat) on success
-*/
-function check_ip_validity(req, res, next) {
-
-	/* get IP */
-    var user_ip = req.connection.remoteAddress;
-    var ip_addr = req.headers["x-forwarded-for"];
-    if (ip_addr) {
-        var list = ip_addr.split(",");
-        user_ip = list[list.length - 1];
-    }
-    
-    /* lookup IP */
-    
-    user_db.find({ip:user_ip})
-    	   .sort({last_post:-1})
-    	   .exec(function (e, d) {
-    	   		if(e) {
-	    	   		res.json({ failure: "session_expiry" });
-	    	   		return;
-    	   		}
-                else if(d[0]) {
-	                if (req.cookies.password_livechan != d[0].session_key) {
-						res.json({ failure: "session_expiry" });
-						return;
-	                } else if ((d[0].last_post.getTime() + 5000) > (new Date).getTime()) {
-	                	var now = new Date;
-		                user_db.update(
-								{ _id: d[0]._id },
-								{ $set: {last_post: now.setTime(now.getTime() + 10000)}},
-								function(){
-							    	if (req.files && req.files.image && req.files.image.path)
-										fs.unlink(req.files.image.path); /* delete file */
-									res.json({ failure: "countdown_violation" });
-									return;
-								});						
-						return;
-					} else if (d[0].banned_rooms.indexOf(req.params.id) > -1) {
-				    	if (req.files && req.files.image && req.files.image.path)
-							fs.unlink(req.files.image.path); /* delete file */
-						res.json({ failure: "ban_violation" });
-						return;
-	                } else {
-	                	var now = new Date;
-	                	user_db.update(
-							{ _id: d[0]._id },
-							{ $set: {last_post: now}},
-							function(){
-								format_image(req, res, next, add_to_chat);
-							});
-	                }
-                } else {
-					if (req.files && req.files.image && req.files.image.path)
-						fs.unlink(req.files.image.path); /* delete file */
-	                res.json({ failure: "session_expiry" });
-	    	   		return;
-                }
-            });
-}
-
-/* format_image
-	- checks for image and sets data accordingly
-	calls generate_thumbnail(req, res, next, data, callback) in case of image/video
-	skips to format_post(req, res, next, data, callback) otherwise
-*/	
-function format_image(req, res, next, callback) {
-  try { /* TODO: remove this try catch */
-	var data = {}; /* to be stored in db and passed to clients */
-
-	/* no image uploaded */
-	
-	if (req.files && req.files.image &&
-		req.files.image.size === 0 || invalid_extension(req.files.image.path)) {
-	
-        fs.unlink(req.files.image.path); /* delete blank file */
-        
-        if (/^\s*$/.test(req.body.body)) {
-        
-            res.json({ failure: "nothing substantial submitted" });
-            return;
-            
-        } else {
-        	format_post(req, res, next, data, callback)
-        }
-    }
-    
-    /* video uploaded */
-    
-    else if (is_video(req.files.image.path)) {
-        var command = "ffprobe -print_format json -show_streams " + req.files.image.path;
-        exec(command, function(err, stdout, stderr) {
-            if (err) {
-                console.log("ffmpeg size error", err);
-                return;
-            }
-            try {
-                var stream_data = JSON.parse(stdout);
-            } catch(e) {
-                console.log("ffmpeg size error", stderr);
-                return;
-            }
-            var video_stream = null;
-            for (var i in stream_data.streams) {
-                if (stream_data.streams[i].codec_type === "video") {
-                    video_stream = stream_data.streams[i];
-                    break;
-                }
-            }
-            if (video_stream !== null && video_stream.width && video_stream.height) {
-                data.image = req.files.image.path;
-                data.image_filename = req.files.image.originalFilename;
-                data.image_filesize = fs.statSync(data.image).size;
-                data.image_width = video_stream.width;
-                data.image_height = video_stream.height;
-                generate_thumbnail(req, res, next, data, callback)
-            } else {
-                console.log("ffmpeg could not find video stream", stderr);
-                return;
-            }
-        });
-    }
-	
-    /* image uploaded */
-    
-    else {
-    	var gm = require('gm').subClass({
-	    imageMagick: useImageMagick // Necessary for hosted version, remove if you're using graphicmagick
-	});
-        gm(req.files.image.path).size(function (err, dimensions) {
-            if (err) {
-                console.log("gm size error", err);
-                return;
-            }
-            
-            if (dimensions.height > 0 && dimensions.width > 0) {
-                data.image = req.files.image.path;
-                data.image_filename = req.files.image.originalFilename;
-                data.image_filesize = fs.statSync(data.image).size;
-                data.image_width = dimensions.width;
-                data.image_height = dimensions.height;
-            }
-			generate_thumbnail(req, res, next, data, callback)
-
-        });
-    }
-
-  } catch(e) {
-	  	if (req.files && req.files.image && req.files.image.path)
-			fs.unlink(req.files.image.path); /* delete blank file */
-		res.json({failure: "illegitimate_request"});
-        return;
-  }
-}
-
-/* generate_thumbnail
-	- generates thumbnail for image
-	calls format_post(req, res, next, data, callback) on completion
-*/
-function generate_thumbnail(req, res, next, data, callback) {
-  try { /* TODO: remove this try catch */
-    var scale = Math.min(250/data.image_width, 100/data.image_height, 1);
-    var thumb_width = Math.round(scale * data.image_width);
-    var thumb_height = Math.round(scale * data.image_height);
-    data.thumb = "public/tmp/thumb/" + data.image.match(/([\w\-]+)\.\w+$/)[1] + ".jpg";
-
-    if (is_video) {
-        var command = "ffmpeg -i "+req.files.image.path+" -s "+thumb_width+"x"+thumb_height+" -vframes 1 "+data.thumb;
-        exec(command, function(err, stdout, stderr) {
-            if (err) {
-                console.log("thumbnail creation error", err);
-                return;
-            }
-            format_post(req, res, next, data, callback);
-        });
-    } else {
-        var gm = require('gm').subClass({
-            imageMagick: useImageMagick // Necessary for hosted version, remove if you're using graphicmagick
-        });
-        gm(data.image)
-            .out("-delete", "1--1") // use first frame only; only needed for ImageMagick
-            .thumb(thumb_width, thumb_height, data.thumb, function(err) {
-                if (err) {
-                    console.log("thumbnail creation error", err);
-                    return;
-                }
-                format_post(req, res, next, data, callback);
-            });
-    }
-  } catch(e) {
-	  	if (req.files && req.files.image && req.files.image.path)
-			fs.unlink(req.files.image.path); /* delete blank file */
-		res.json({failure: "illegitimate_request"});
-        return;
-  }
-}
-
-/* format_post:
-	- checks for text violations
-	- generates tripcode
-	calls callback(data) on success
-*/
-function format_post(req, res, next, data, callback) {	
-  try { /* TODO: remove this try catch */
-    /* length checks */
-    if (!req.body) {
-	   if (req.files && req.files.image && req.files.image.path)
-    		fs.unlink(req.files.image.path); /* delete blank file */
-		res.json({failure: "illegitimate_request"});
-        return;
-    }
-    
-    if (req.body.body && req.body.body.length > 2000) {
-        req.body.body = req.body.body.substring(0, 1999) + "...";
-    }
-
-    if (req.body.body.split("\n").length > 100) {
-        req.body.body = req.body.body.split("\n", 100).join("\n");
-    }
-
-	/* generate tripcode */
-	
-    var trip_index = req.body.name.indexOf("#");
-	var trip;
-	
-    if (trip_index > -1) {
-        var trip = req.body.name.substr(trip_index + 1);
-        var secure = trip.indexOf("#") === 0;
-        if (secure) {
-            trip = crypto.createHash('sha1').update(trip.substr(1) + 
-            		securetrip_salt).digest('base64').toString();
-        }
-
-        data.trip = (secure ? "!!" : "!") + tripcode(trip);
-        req.body.name = req.body.name.slice(0, trip_index);
-    }
-    
-    /* truncate name */
-
-    if (req.body.name.length > 40) {
-        req.body.name = req.body.name.substring(0, 39) + "...";
-    }
-
-    if (req.body.convo) {
-        if (req.body.convo.length > 40) {
-            req.body.convo = req.body.convo.substring(0, 39) + "...";
-        }
-    } else {
-        req.body.convo = 'General'; /* default conversation */
-    }
-
-
-    
-    /* everything looks good, we can safely add this chat */
-    
-    count++; /* counter has been promoted */
-
-    data.count = count;
-    data.convo = req.body.convo;
-	data.body = req.body.body;
-    data.name = req.body.name || "Anonymous";
-    data.date = (new Date()).toString();
-    var user_ip = req.connection.remoteAddress;
-    var ip_addr = req.headers["x-forwarded-for"];
-    if (ip_addr) {
-        var list = ip_addr.split(",");
-        user_ip = list[list.length - 1];
-    }
-    data.ip = user_ip;
-    data.chat = req.params.id;
-    
-    
-    /* determine if OP of conversation topic */
-    if (data.convo && data.convo !== "General") {
-        chat_db.findOne({
-            convo: data.convo,
-            chat: data.chat,
-            is_convo_op: true
-        }).exec(function (err, convo_ent) {
-            if (!convo_ent) {
-                data.is_convo_op = true;
-                data.convo_id = data.count;
-                /* cool down increase */
-            } else {
-                data.is_convo_op = false;
-                data.convo_id = convo_ent.count;
-            }
-            callback(data);
-        });
-    } else {
-        callback(data);
-    }
-
-
-	/* give the client information about the post */
-    res.json({
-        success: "success_posting",
-        id: data.count
-    });
-    
-    return;
-  } catch(e) {
-	  	if (req.files && req.files.image && req.files.image.path)
-			fs.unlink(req.files.image.path); /* delete blank file */
-		res.json({failure: "illegitimate_request"});
-        return;
-  }
-}
-
-/* REQUESTS */
-
-app.all('*',function(req,res,next){
-  if(!req.connection.encrypted && secure_server != server)
-    res.redirect('https://livechan.net'+req.url);
-  else
-    next();
-});
-
-app.get('/login', function (req, res) {
-    "use strict";
-    res.send('<html><head><meta name="viewport" content="width=device-width,user-scalable=no"><link type="text/css" rel="stylesheet" href="style.css"></head><body><div class="center container"><div>Please prove you aren\'t a robot</div><br/><img src="/captcha.jpg"/><form action="/login" method="post"><br/><input type="text" name="digits"/><input type="hidden" name="page" value="' + req.query.page + '"/></form></div></body></html>');
-});
-
-app.post('/login', function (req, res) {
-    "use strict";
-    res.type("text/plain");
-    var user_ip = req.connection.remoteAddress;
-    var ip_addr = req.headers["x-forwarded-for"];
-    if (ip_addr) {
-        var list = ip_addr.split(",");
-        user_ip = list[list.length - 1];
-    }
-
-	var secure_text = crypto.createHash('sha1').update(req.body.digits + securetrip_salt).digest('base64').toString();
-	user_db.find({captcha:req.session.captcha}).exec(function(e,d){
-		if(d.length > 0) {
-			res.json({
-                failure: "captcha_used"
-            });
-		} else {
-			if (secure_text === req.session.captcha) {
-		        var key = (Math.random() * 1e17).toString(36);
-		        var info = req.headers['user-agent'] + user_ip + key;
-		        var password = crypto.createHash('sha1').update(info).digest('base64').toString();
-		        res.cookie('password_livechan', password, {
-		            maxAge: 86400000,
-		            httpOnly: false
-		        });
-		
-		        if (req.body.page) {
-		            res.redirect(req.body.page);
-		        } else {
-		            res.json({
-		                success: "captcha"
-		            });
-		        }
-				
-				var now = new Date;
-		
-		        var data = {
-		            session_key: password,
-		            captcha: req.session.captcha,
-		            ip: user_ip,
-		            last_post: now.setTime(now.getTime() - 6000)
-		        };
-		
-		        new user_db(data).save(function () {
-		            /*user_db.find().exec(function (e, d) {
-		                console.log("session found", e, d);
-		            });*/
-		        });
-		
-		        return;
-		    }
-			if (req.body.page) {
-			        res.send("You mistyped the captcha!");
-		    } else {
-		        res.json({
-		            failure: "You mistyped the captcha."
-		        });
-		    }
-		}
-	})
-    
-});
-
-app.get('/', function (req, res) {
-    "use strict";
-    res.redirect('/chat/home');
-});
-
-app.get('/chat/:id([a-z0-9]+)', function (req, res) {
-    "use strict";
-    if (boards.indexOf(req.params.id) < 0 && req.params.id !== "home") {
-        res.send("Board doesn't exist :(");
-        return;
-    }
-    res.sendfile('pages/index.html');
-    return;
-});
-
-app.get('/data:ops((?:_convo)?)/:id([a-z0-9]+)', function (req, res) {
-    "use strict";
-    if (req.params.id !== "all" && boards.indexOf(req.params.id) < 0) {
-        res.send("Does not exist :(");
-        return;
-    }
-    var search = {};
-    var limit = 0;
-    var fields = "";
-    if (req.params.id === "all") {
-        limit = 30;
-        fields = all_fields;
-    } else {
-        search.chat = req.params.id;
-        limit = 150;
-        fields = board_fields;
-    }
-    if (req.params.ops === "_convo") {
-        search.is_convo_op = true;
-        limit = 30;
-    }
-    chat_db.find(search)
-        .sort({
-            count: -1
-        })
-        .select(fields)
-        .limit(limit)
-        .exec(function (e, d) {
-            if (!e) {
-                res.json(d);
-            } else {
-                res.send('db_error');
-            }
-        });
-});
-
-app.get('/ban/:password([a-z0-9]+)/:id([0-9]+)/:board', function (req, res, next) {
-    "use strict";
-    var chat_id = req.params.id;
-    var password = req.params.password;
-    var type = req.params.text;
-    var board = req.params.board;
-
-	var hash_pass = crypto.createHash('sha1').update(password).digest('base64');
-	
-	if (hash_pass!="Nqesm9E+3GXfOG0KgJq8YmizCho="){
-		console.log("ATTEMPTED PASS", password);
-		res.json({failure:"wrong password"});
-		return;
-	}
-		
-	chat_db.find({count:chat_id},
-		function(e,d){
-			if(d[0] && d[0].ip) {
-				user_db.update({ip:d[0].ip},{
-					$push:{
-						banned_rooms:board
-					}
-				}, {multi:true}, function(err){
-					if(!err) {
-						res.json({success:"banned "+d[0].ip+" from "+board});
-					} else {
-						res.json({failure:"couldn't ban "+d[0].ip+" from "+board});
-					}
-				});
-			} else {
-				res.json({failure:"couldn't find ip"});
-			}
-		});
-	return;
-
-});
-
-app.get('/delete/:password([a-z0-9]+)/:id([0-9]+)', function (req, res, next) {
-    "use strict";
-    var chat_id = req.params.id;
-    var password = req.params.password;
-
-	var hash_pass = crypto.createHash('sha1').update(password).digest('base64');
-	
-	if (hash_pass!="Nqesm9E+3GXfOG0KgJq8YmizCho="){
-		console.log("ATTEMPTED PASS", password);
-		res.json({failure:"wrong password"});
-		return;
-	}
-	
-	chat_db.update(
-    	{count:chat_id},
-    	{$set:
-    		{body:"deleted",
-    		convo:"General",
-    		name:"deleted",
-    		image_filename:""}
-    	},function(err){
-	    	if(!err) {
-		    	chat_db.find({count:chat_id},
-		    		function(e,d){
-			    		if(d[0] && d[0].image) {
-				    		fs.unlink(d[0].image);
-				    		fs.unlink(d[0].thumb);
-			    		}
-			    		d[0].image = "";
-						add_to_chat(d[0]);
-					})
-	    	}
-    	});
-    
-    res.json({success:"deleted "+chat_id});
-    return;
-    
-});
-
-app.get('/set/:password([a-z0-9]+)/:id([0-9]+)/:text', function (req, res, next) {
-    "use strict";
-    var chat_id = req.params.id;
-    var password = req.params.password;
-    var text = decodeURI(req.params.text);
-
-	var hash_pass = crypto.createHash('sha1').update(password).digest('base64');
-	
-	if (hash_pass!="Nqesm9E+3GXfOG0KgJq8YmizCho="){
-		console.log("ATTEMPTED PASS", password);
-		res.json({failure:"wrong password"});
-		return;
-	}
-	
-	chat_db.update(
-    	{count:chat_id},
-    	{$set:
-    		{body: text}
-    	},function(err){
-	    	if(!err) {
-		    	chat_db.find({count:chat_id},
-		    		function(e,d){
-						add_to_chat(d[0]);
-					})
-	    	}
-    	});
-    
-    res.json({success:"reset "+chat_id});
-    return;
-    
-});
-
-app.post('/chat/:id([a-z0-9]+)', function (req, res, next) {
-    "use strict";
-    res.type("text/plain");
-    if (req.params.id !== "all" && boards.indexOf(req.params.id) < 0) {
-    	if (req.files && req.files.image && req.files.image.path)
-    		fs.unlink(req.files.image.path); /* delete blank file */
-        res.json({failure: "This board does not exist."});
-        return;
-    }
-    check_ip_validity(req, res, next);
-    return;
-});
-
-/* 404 */
-app.get('*', function(req, res){
-    res.status(404).sendfile('pages/404.html');
-});
-
-/* 404 */
-app.post('*', function(req, res){
-    res.status(404).sendfile('pages/404.html');
-});
-
-/* reduce logging during production runs */
-if (port==80)
-	io.set('log level', 1);
-
-/* join a chat room */
-io.sockets.on('connection', function (socket) {
-    "use strict";
-    socket.emit('request_location', "please return a chat room");
-    socket.on('subscribe', function (data) {
-        socket.join(data);
-    });
-    socket.on('unsubscribe', function (data) {
-    	socket.leave(data);
-    });
-});
